modified new file paths related to actions
