+++ bc-react/src/routes/VotePageMockUp.js
@@ -2,7 +2,7 @@
 import React, {Component} from 'react';
 import SideBar from '../components/SideBar';
 import SideBarMini from '../components/SideBarMini';
-import {updateUser} from '../actions';
+import {updateUser} from '../actions/UserActions';
 import userStore from '../stores/UserStore';
 import Header from '../components/Header';
